src/pdf/pyproject.toml
[project]
name = "ekip-enedis-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fitz>=0.0.1.dev2",
    "numpy>=2.2.2",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.1.0",
    "pymupdf>=1.25.2",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
]
